+++ src/main/java/net/tracystacktrace/stackem/impl/DeepMeta.java
@@ -1,12 +1,18 @@
 package net.tracystacktrace.stackem.impl;
 
+import it.unimi.dsi.fastutil.objects.Object2ObjectArrayMap;
+import net.tracystacktrace.stackem.modifications.entityvariation.EntityVariation;
 import net.tracystacktrace.stackem.modifications.moon.CelestialMeta;
 import net.tracystacktrace.stackem.modifications.moon.EnumCelestialCycle;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.*;
 
 public class DeepMeta {
     //can be null-ed when flushed
     public CelestialMeta moonData;
     public CelestialMeta sunData;
+    public Map<String, List<String>> entityVariations = new HashMap<>();
 
     /* moon data related stuff */
 
@@ -36,10 +42,33 @@ public class DeepMeta {
         this.sunData = metadata;
     }
 
+    /* entity variations system */
+
+    public void addEntityVariation(@NotNull EntityVariation.Description description) {
+        if (this.entityVariations == null) {
+            this.entityVariations = new Object2ObjectArrayMap<>();
+        }
+        if (!this.entityVariations.containsKey(description.entity())) {
+            this.entityVariations.put(description.entity(), new ArrayList<>());
+        }
+        if (description.forceOnly()) {
+            this.entityVariations.get(description.entity()).clear();
+        }
+
+        this.entityVariations.get(description.entity()).addAll(Arrays.asList(description.variants()));
+    }
+
+    public boolean hasVariationsForEntity(@NotNull String entityName) {
+        if (this.entityVariations == null)
+            return false;
+        return this.entityVariations.containsKey(entityName);
+    }
+
     /* general methods */
 
     public void flush() {
         this.moonData = null;
         this.sunData = null;
+        this.entityVariations = null;
     }
 }